refactor: mvc모델 역할별로 패키지 생성
