+++ src/main/java/lotto/domain/LottoFactory.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
