+++ src/main/java/lotto/domain/Money.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Money {
     private final int PRICE_OF_LOTTO = 1_000;
