+++ src/main/java/lotto/domain/Result.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.EnumMap;
 import java.util.Map;
