+++ src/test/java/lotto/MoneyTest.java
@@ -2,6 +2,7 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
